Migrate alias-adapters from src/utils to .dev/utils
