+++ .dev/utils/alias-adapters/vite-adapter.mjs
@@ -1,7 +1,7 @@
 /**
- * @file vitest-adapter.mjs
- * @description Adapter for validating Vitest configuration imports from alias.config.mjs
- * @path src/utils/alias-adapters/vitest-adapter.mjs
+ * @file vite-adapter.mjs
+ * @description Adapter for validating Vite configuration imports from alias.config.mjs
+ * @path .dev/utils/alias-adapters/vite-adapter.mjs
  */
 
 import { resolve } from 'path';
@@ -9,53 +9,53 @@ import { readFileSync } from 'fs';
 import { BaseConfigAdapter } from './base-adapter.mjs';
 
 /**
- * Adapter for Vitest configuration files (vitest.config.mjs).
+ * Adapter for Vite configuration files (vite.config.mjs).
  * Validates that the config correctly imports from alias.config.mjs.
- * Does NOT support write operations (vitest.config.mjs should always import from alias.config.mjs).
+ * Does NOT support write operations (vite.config.mjs should always import from alias.config.mjs).
  *
- * @class VitestConfigAdapter
+ * @class ViteConfigAdapter
  * @extends BaseConfigAdapter
  */
-export class VitestConfigAdapter extends BaseConfigAdapter {
+export class ViteConfigAdapter extends BaseConfigAdapter {
   #filePath;
 
   /**
-   * Creates VitestConfigAdapter instance.
+   * Creates ViteConfigAdapter instance.
    *
    * @param {string} [projectRoot=process.cwd()] - Project root directory
    */
   constructor(projectRoot = process.cwd()) {
     super();
-    this.#filePath = resolve(projectRoot, 'vitest.config.mjs');
+    this.#filePath = resolve(projectRoot, 'vite.config.mjs');
   }
 
   /**
    * Returns format identifier.
    *
-   * @returns {string} 'vitest'
+   * @returns {string} 'vite'
    */
   getFormat() {
-    return 'vitest';
+    return 'vite';
   }
 
   /**
    * Returns file paths this adapter handles.
    *
-   * @returns {string[]} Array containing vitest.config.mjs path
+   * @returns {string[]} Array containing vite.config.mjs path
    */
   getFilePaths() {
     return [this.#filePath];
   }
 
   /**
-   * Validates that vitest.config.mjs correctly imports from alias.config.mjs.
-   * Returns empty object since Vitest doesn't need explicit aliases (uses import).
+   * Validates that vite.config.mjs correctly imports from alias.config.mjs.
+   * Returns empty object since Vite doesn't need explicit aliases (uses import).
    *
-   * @returns {Promise<Object>} Empty object (Vitest imports from alias.config.mjs)
+   * @returns {Promise<Object>} Empty object (Vite imports from alias.config.mjs)
    * @throws {Error} If file doesn't import from alias.config.mjs
    *
    * @example
-   * const adapter = new VitestConfigAdapter();
+   * const adapter = new ViteConfigAdapter();
    * const aliases = await adapter.read(); // Returns: {}
    */
   async read() {
@@ -70,11 +70,11 @@ export class VitestConfigAdapter extends BaseConfigAdapter {
 
       if (!hasImport || !hasAliasEntries) {
         throw new Error(
-          `[OMH] vitest.config.mjs must import aliasEntries from './alias.config.mjs'`
+          `[OMH] vite.config.mjs must import aliasEntries from './alias.config.mjs'`
         );
       }
 
-      // Return empty object - Vitest uses import directly
+      // Return empty object - Vite uses import directly
       return {};
     } catch (error) {
       if (error.message.includes('[OMH]')) {
@@ -87,22 +87,22 @@ export class VitestConfigAdapter extends BaseConfigAdapter {
   }
 
   /**
-   * Write operation not supported for Vitest config.
-   * Vitest should always import from alias.config.mjs.
+   * Write operation not supported for Vite config.
+   * Vite should always import from alias.config.mjs.
    *
    * @throws {Error} Always throws - write not supported
    */
   async write(_aliases) {
     throw new Error(
-      `[OMH] vitest.config.mjs should import from alias.config.mjs - manual write not supported`
+      `[OMH] vite.config.mjs should import from alias.config.mjs - manual write not supported`
     );
   }
 
   /**
-   * Validates that Vitest config correctly imports from alias.config.mjs.
+   * Validates that Vite config correctly imports from alias.config.mjs.
    * Always returns valid=true if import is present, valid=false otherwise.
    *
-   * @param {Object} _expected - Expected aliases (ignored for Vitest)
+   * @param {Object} _expected - Expected aliases (ignored for Vite)
    * @returns {Promise<Object>} ValidationResult
    */
   async validate(_expected) {
