+++ tests/unit/adapters/base-adapter.unit.test.mjs
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect } from 'vitest';
-import { BaseConfigAdapter } from '#/utils/alias-adapters/base-adapter.mjs';
+import { BaseConfigAdapter } from '../../../.dev/utils/alias-adapters/base-adapter.mjs';
 
 // Test implementation of BaseConfigAdapter
 class TestAdapter extends BaseConfigAdapter {
